Moved CenteredSeries mixin as utilities to Series folder.
